Added swagger and refactored docker
