+++ .gitignore
@@ -6,3 +6,4 @@ build/
 htmlCoverage/
 __pycache__/
 .pytest_cache
+.env
